+++ ggc-core/src/ggc/BreakdownTransaction.java
@@ -1,11 +1,11 @@
 package ggc;
 
 public class BreakdownTransaction extends Transaction{
-    private float _paymentPrice;
+    private double _paymentPrice;
     private int _dateDeadline;
     private String _recipe;
 
-    public BreakdownTransaction(int id, int paymentDate, String productId, String partnerId, int quantity, int price, String recipe){       
+    public BreakdownTransaction(int id, int paymentDate, String productId, String partnerId, int quantity, double price, String recipe){
         super(id, paymentDate, productId, partnerId, quantity, price);
         _recipe = recipe; 
     }
